requirements.txt
requests
numpy
pandas
seaborn
flask
pymongo
fastapi
pydantic
uvicorn
python-dotenv